requirements.txt
# Core FastAPI
fastapi==0.104.1
uvicorn==0.24.0
starlette==0.27.0

# Pydantic for validation
pydantic==2.5.0
pydantic-settings==2.1.0

# OpenAI Integration
openai==1.3.0

# GitHub Integration
PyGithub==2.1.1
gitpython==3.1.40

# HTTP Requests
requests==2.31.0
httpx==0.25.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Environment management
python-dotenv==1.0.0
pyyaml==6.0.1

# Testing & Evaluation
playwright==1.40.0
pytest==7.4.3
pytest-asyncio==0.21.1

# Utilities
python-dateutil==2.8.2
aiofiles==23.2.1
python-multipart==0.0.6

# Security
python-jose==3.3.0
passlib==1.7.4
cryptography==41.0.7

# Async support
aiohttp==3.9.1

# Data processing (optional)
pandas==2.1.3
numpy==1.24.3
